+++ src/lib/FrodoLib.ts
@@ -33,6 +33,9 @@ import EnvFederationEnforcementOps, {
 } from '../ops/cloud/EnvFederationEnforcementOps';
 import EnvPromotionOps, { EnvPromotion } from '../ops/cloud/EnvPromotionOps';
 import EnvReleaseOps, { EnvRelease } from '../ops/cloud/EnvReleaseOps';
+import EnvServiceAccountScopesOps, {
+  EnvServiceAccountScopes,
+} from '../ops/cloud/EnvServiceAccountScopesOps';
 import EnvSSOCookieConfigOps, {
   EnvSSOCookieConfig,
 } from '../ops/cloud/EnvSSOCookieConfigOps';
@@ -128,6 +131,7 @@ export type Frodo = {
       EnvCustomDomains &
       EnvFederationEnforcement &
       EnvRelease &
+      EnvServiceAccountScopes &
       EnvSSOCookieConfig & {
         cert: EnvCertificate;
         csr: EnvCSR;
@@ -302,6 +306,7 @@ const FrodoLib = (config: StateInterface = {}): Frodo => {
         ...EnvCustomDomainsOps(state),
         ...EnvFederationEnforcementOps(state),
         ...EnvReleaseOps(state),
+        ...EnvServiceAccountScopesOps(state),
         ...EnvSSOCookieConfigOps(state),
         cert: EnvCertificatesOps(state),
         csr: EnvCSRsOps(state),